requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
Pillow>=10.0
face-recognition>=1.3
opencv-python>=4.8
numpy>=1.24
psycopg2-binary>=2.9
python-dotenv>=1.0
openpyxl>=3.1.2
